Update Environment.cpp
It is supposed to update the matrix when the control is not equal to zero, which is used to mean that the agent is supposed to stay still and do not move any steps. Thus the value of the matrix should not be updated to the value of the new position because the agent do not ever move.